Removed the signed/unsigned mismatch compiler warnings

+++ bitstream.c
@@ -138,7 +138,7 @@ drain_into_ancillary(lame_global_flags *gfp,int remainingBits)
       const char * version;
       version = get_lame_version();
       if (remainingBits >= 32) 
-	for (i=0; i<strlen(version) && remainingBits >=8 ; ++i) {
+	for (i=0; i<(int)strlen(version) && remainingBits >=8 ; ++i) {
 	  remainingBits -= 8;
 	  putbits2(gfp,(unsigned int)version[i],8);
 	}
@@ -408,7 +408,7 @@ encodeSideInfo2(lame_global_flags *gfp,int bitsPerFrame)
 
     {
 	int old = gfc->h_ptr;
-	assert(gfc->header[old].ptr == gfc->sideinfo_len * 8);
+	assert(gfc->header[old].ptr == (int)(gfc->sideinfo_len * 8));
 
 	gfc->h_ptr = (old + 1) & (MAX_HEADER_BUF - 1);
 	gfc->header[gfc->h_ptr].write_timing =
@@ -528,7 +528,7 @@ HuffmanCode(lame_global_flags *gfp, unsigned int table_select, int x, int y)
 	/* use ESC-words */
 	if (x > 14) {
 	    int linbitsx = x - 15;
-	    assert(linbitsx <= h->linmax);
+	    assert(linbitsx <= (int)h->linmax);
 	    ext |= linbitsx << 1;
 	    xbits = linbits;
 	    x = 15;
@@ -536,7 +536,7 @@ HuffmanCode(lame_global_flags *gfp, unsigned int table_select, int x, int y)
 
 	if (y > 14) {
 	    int linbitsy = y - 15;
-	    assert(linbitsy <= h->linmax);
+	    assert(linbitsy <= (int)h->linmax);
 	    ext <<= linbits;
 	    ext |= linbitsy;
 	    xbits += linbits;
@@ -605,7 +605,7 @@ ShortHuffmancodebits(lame_global_flags *gfp,int *ix, gr_info *gi)
     int region1Start;
     
     region1Start=3*gfc->scalefac_band.s[3];
-    if (region1Start > gi->big_values) 	region1Start = gi->big_values;
+    if (region1Start > (int)gi->big_values) 	region1Start = gi->big_values;
 
     /* short blocks do not have a region2 */
     bits  = Huffmancodebits(gfp,gi->table_select[0], 0, region1Start, ix);
@@ -702,8 +702,8 @@ writeMainData(lame_global_flags *gfp,
 		DEBUGF("<%ld> ", gfc->bs.totbit-hogege);
 #endif
 		/* does bitcount in quantize.c agree with actual bit count?*/
-		assert(data_bits==gi->part2_3_length-gi->part2_length);
-		assert(scale_bits==gi->part2_length);
+		assert(data_bits==(int)gi->part2_3_length-(int)gi->part2_length);
+		assert(scale_bits==(int)gi->part2_length);
 		tot_bits += scale_bits + data_bits;
 
 	    } /* for ch */
@@ -746,8 +746,8 @@ writeMainData(lame_global_flags *gfp,
 	    data_bits +=huffman_coder_count1(gfp,l3_enc[gr][ch], gi);
 
 	    /* does bitcount in quantize.c agree with actual bit count?*/
-	    assert(data_bits==gi->part2_3_length-gi->part2_length);
-	    assert(scale_bits==gi->part2_length);
+	    assert(data_bits==(int)gi->part2_3_length-(int)gi->part2_length);
+	    assert(scale_bits==(int)gi->part2_length);
 	    tot_bits += scale_bits + data_bits;
 	} /* for ch */
     } /* for gf */
